+++ src/App.js
@@ -2,8 +2,8 @@ import React, { Component} from 'react';
 import './App.css';
 import Button from 'react-bootstrap/Button';
 import DriverHome from './Driver/DriverHome'
-import {RegistroVehiculo} from './RegistroVehiculo';
-
+import {RegistroVehiculo} from './RegistroVehiculo/RegistroVehiculo';
+import RegistroUsuario from "./RegistroUsuario/RegistroUsuario";
 import {Route, BrowserRouter as Router} from 'react-router-dom';
 
 class App extends Component {
@@ -21,6 +21,7 @@ class App extends Component {
                       <Route exact path="/" component={DriverHome}/>
                       <Route path="/driver" component={DriverHome}/>
                       <Route path="/registerVehic" component={RegistroVehiculo}/>
+                      <Route path="/registerUsuario" component={RegistroUsuario} />
                       
                   </div>
               </Router>
